+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## 32.1.3
+### Enhancements
+- **Module Improvements**:
+  - Updated the message argument to obj_message in avi_albservicesfiledownload
 
 ### Bug Fixes
 - avi_serviceenginegroup_advanced Module is removed. All fields previously only available through avi_serviceenginegroup_advanced are now available directly on vmware.alb.avi_serviceenginegroup. Action required: Playbooks and roles using avi_serviceenginegroup_advanced must be updated to use avi_serviceenginegroup instead, passing all fields (both previously "basic" and "advanced") to that single module.